+++ src/controller/GameController.js
@@ -1,20 +1,18 @@
 import GameInput from "../view/GameInput.js";
 import GameOutput from "../view/GameOutput.js";
-import NumberValidate from "../validate/NumberValidate.js";
-import GetNumber from "../model/GetNumber.js";
+import Exception from "../util/Exception.js";
+import GetNumber from "../util/GetNumber.js";
 import GameResult from "../model/GameResult.js";
 import { LOTTO_DATA } from "../constant/Data.js";
 
 class GameController {
   #gameInput;
   #gameOutput;
-  #getNumber;
   #gameResult;
 
   constructor() {
     this.#gameInput = new GameInput();
     this.#gameOutput = new GameOutput();
-    this.#getNumber = new GetNumber();
     this.#gameResult = new GameResult();
   }
 
@@ -22,7 +20,7 @@ class GameController {
     while (true) {
       try {
         const money = await this.#gameInput.readPurchaseMoney();
-        this.#purchaseMoneyValidate(money);
+        Exception.purchaseMoneyValidate(money);
         return money;
       } catch (error) {
         this.#gameOutput.printErrorMesssage(error);
@@ -36,25 +34,19 @@ class GameController {
     return new_lotto;
   }
 
-  #purchaseMoneyValidate(money) {
-    NumberValidate.validateNonNumber(money);
-    NumberValidate.validateSmallNumber(money);
-    NumberValidate.validateDivideThousand(money);
-  }
-
   #purchaseLotto(money) {
     return parseInt(money / LOTTO_DATA.lottoPrice);
   }
 
   #newLotto(purchase_lotto) {
-    return Array.from({ length: purchase_lotto }, () => this.#getNumber.purchaseLotto());
+    return Array.from({ length: purchase_lotto }, () => GetNumber.purchaseLotto());
   }
 
   async #winningLotto() {
     while (true) {
       try {
         const winning_lotto = await this.#gameInput.readWinningLotto();
-        return this.#getNumber.winningLotto(winning_lotto);
+        return GetNumber.winningLotto(winning_lotto);
       } catch (error) {
         this.#gameOutput.printErrorMesssage(error);
       }
@@ -65,18 +57,12 @@ class GameController {
     while (true) {
       try {
         const bonus = await this.#gameInput.readBonusNumber();
-        this.#bonusNumberValidate(bonus, winning_lotto);
+        Exception.bonusNumberValidate(bonus, winning_lotto);
         return bonus;
       } catch (error) {
         this.#gameOutput.printErrorMesssage(error);
       }
     }
-  }
-
-  #bonusNumberValidate(bonus, winning_lotto) {
-    NumberValidate.validateNonNumber(bonus);
-    NumberValidate.validateBonusDup(bonus, winning_lotto);
-    NumberValidate.validateBonusRange(bonus);
   }
 
   async startGame() {
